Added code for printing simulation in SJN
(Draft) I think code doesn't look that good and can still be polished. Still has a known bug when all of the jobs' arrival time is the same. Code needs more testing because I doubt that there are still more bugs. All of the changes are inside the void SJN() function.